+++ diploma_thesis/scheduling_rules/cr.py
@@ -4,6 +4,7 @@ import torch
 from .scheduling_rule import SchedulingRule, MachineState, WaitInfo
 from environment.job import Job
 
+# TODO: Pass Reduction strategy
 
 class CRSchedulingRule(SchedulingRule):
     """
@@ -12,10 +13,10 @@ class CRSchedulingRule(SchedulingRule):
     """
 
     def __call__(self, machine_state: MachineState) -> Job | WaitInfo:
-        remaining_processing_times = torch.LongTensor(
-            [job.remaining_processing_time for job in machine_state.queue]
+        remaining_processing_times = torch.FloatTensor(
+            [job.remaining_processing_time() for job in machine_state.queue]
         )
-        due_times = torch.LongTensor(
+        due_times = torch.FloatTensor(
             [job.time_until_due(machine_state.now) for job in machine_state.queue]
         )
