+++ diploma_thesis/scheduling_rules/lro.py
@@ -11,7 +11,7 @@ class LROSchedulingRule(SchedulingRule):
     """
 
     def __call__(self, machine_state: MachineState) -> Job | WaitInfo:
-        values = [job.remaining_operations_count for job in machine_state.queue]
+        values = torch.FloatTensor([job.remaining_operations_count for job in machine_state.queue])
         idx = torch.argmin(values)
 
         return machine_state.queue[idx]
