+++ diploma_thesis/scheduling_rules/lwkr.py
@@ -11,7 +11,7 @@ class LWRKSchedulingRule(SchedulingRule):
     """
 
     def __call__(self, machine_state: MachineState) -> Job | WaitInfo:
-        values = [job.remaining_processing_time for job in machine_state.queue]
+        values = torch.FloatTensor([job.remaining_processing_time() for job in machine_state.queue])
         idx = torch.argmin(values)
 
         return machine_state.queue[idx]
