+++ diploma_thesis/scheduling_rules/npt.py
@@ -13,7 +13,7 @@ class NPTSchedulingRule(SchedulingRule):
     """
 
     def __call__(self, machine_state: MachineState) -> Job | WaitInfo:
-        values = [job.next_operation_processing_time for job in machine_state.queue]
+        values = torch.FloatTensor([job.next_operation_processing_time() for job in machine_state.queue])
         idx = torch.argmin(values)
 
         return machine_state.queue[idx]
